Commit the outline-extractor spec and the languages folder
The study and runbook land at plugins/horos/docs/ts-outline/. The map verb
now dispatches through a suffix registry at scripts/languages/, one folder
per language, and the Python mapper moves to languages/python/python.py
with its pinned fixture output unchanged. An unregistered suffix is refused
naming the supported list.

Wildcat-Origin: shoggoth
