+++ plugins/horos/skills/horos/scripts/languages/__init__.py
@@ -1,0 +1,16 @@
+"""The language registry Horos's map verb dispatches through.
+
+One folder per language. An extractor exposes outline(path, source, out)
+and returns an exit code; the registry maps file suffixes onto it. A suffix
+absent here is refused by map, never guessed at.
+"""
+
+from .python import python
+
+EXTRACTORS = {
+    ".py": python.outline,
+}
+
+
+def supported():
+    return ", ".join(sorted(EXTRACTORS))
